use portable type kinds to avoid unwanted compiler behaviour

+++ src/append_funcs.f90
@@ -12,7 +12,7 @@ module Class_append_funcs
         real(dp), allocatable ::  down_collect(:,:)
         real(dp), allocatable ::  spins_collect(:,:)
         integer, allocatable ::  sample_idx(:)
-        integer              ::  me,me_sample,color
+        integer(int64)              ::  me,me_sample,color
         character(len=300)   :: prefix
         type(units)          :: units
     contains
@@ -30,7 +30,7 @@ module Class_append_funcs
             implicit none
             type(collect_quantities) :: self
             type(CFG_t)              :: cfg
-            integer                  :: ierr,color
+            integer(int64)                  :: ierr,color
             type(MPI_Comm)           :: sample_comm
             character(len=300)       :: prefix
     
@@ -79,7 +79,7 @@ module Class_append_funcs
             use mpi_f08
             implicit none
             class(collect_quantities)           :: self
-            integer                             :: i
+            integer(int64)                             :: i
             integer,allocatable                 :: isize(:)
             real(dp), intent(in)                 :: spins(:,:)
             allocate(isize(2))
@@ -185,7 +185,7 @@ module Class_append_funcs
 
         subroutine add_to_arr1D_int(list, element)
             implicit none
-            integer                             :: i,isize
+            integer(int64)                             :: i,isize
             integer, intent(in)                 :: element
             integer, allocatable, intent(inout) :: list(:)
             integer, allocatable                :: clist(:)
@@ -211,7 +211,7 @@ module Class_append_funcs
         subroutine add_to_arr1D_real(list, element)
             !https://stackoverflow.com/questions/28048508/how-to-add-new-element-to-dynamical-array-in-fortran-90
             implicit none
-            integer                             :: i,isize
+            integer(int64)                             :: i,isize
             real(dp), intent(in)                 :: element
             real(dp), allocatable, intent(inout) :: list(:)
             real(dp), allocatable                :: clist(:)
@@ -237,7 +237,7 @@ module Class_append_funcs
         subroutine add_to_arr2D_real(list, element)
             !https://stackoverflow.com/questions/28048508/how-to-add-new-element-to-dynamical-array-in-fortran-90
             implicit none
-            integer                             :: i
+            integer(int64)                             :: i
             integer,allocatable                 :: isize(:)
             real(dp)             , intent(in)    :: element(:)
             real(dp), allocatable, intent(inout) :: list(:,:)
@@ -268,7 +268,7 @@ module Class_append_funcs
     subroutine add_2D_to_arr2D_real(list, element)
         !https://stackoverflow.com/questions/28048508/how-to-add-new-element-to-dynamical-array-in-fortran-90
         implicit none
-        integer                             :: i
+        integer(int64)                             :: i
         integer,allocatable                 :: isize(:),esize(:)
         real(dp)             , intent(in)    :: element(:,:)
         real(dp), allocatable, intent(inout) :: list(:,:)
